+++ NEWS.md
@@ -5,7 +5,7 @@ Version 1.4.6
 * Added 'fac.center' argument in ordiplot to plot canonical coefficients of binary variables as points
 * Added a simple plotting function for the gllvm summary
 * Improved scaling for ordiplot with quadratic model and with biplot = FALSE
-* optima.gllvm now correct for the (previously) implement diagonal extraction with num.lv
+* optima.gllvm and tolerances.gllvm for num.lv now correctly provide tolerances w.r.t. the scaled LV
 
 ## Bug Fixes
 * Bug fixed that prevented increasing he point size of sites in ordiplot with symbols = TRUE